Fix `CreateButton` not being re-translated when language change at runtime
The `CreateButton` is using an optimization to re-render only a certain set
of keys change. However, the `translate` prop was not part of that. This
was causing the component to not re-render when the translate function was
updated due to languague having changed.

This is now fixed by specifying the `translate` as being a prop change
that causes a re-render.

Fixes #2830